This is a little about the project
So in this project we were going to create a nav bar that uses @media queries I was having problems because I did not use the media queries for the bigger size page so when it came to making the links disappear I was having a hard time. Then we went over the solution and I learned that display: none; would let you pull items on and off the page with @media queries.